pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "size-reducer-cli"
version = "0.1.0"
description = "A lightweight CLI tool to reduce image and PDF file sizes."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
  { name = "Arpit Seth" }
]
dependencies = [
  "Pillow>=10.0.0",
  "pillow-heif>=1.0.0",
  "pypdf>=6.0.0"
]

[project.scripts]
size-reducer = "size_reducer_cli.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
